requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.5.0
asttokens==2.4.1
boto3==1.35.10
botocore==1.35.10
click==8.1.7
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
stack-data==0.6.3
starlette==0.38.4
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
watchdog==5.0.1
wcwidth==0.2.13
xmltodict==0.13.0
yq==3.4.3
